requirements.txt
python-telegram-bot 
google-generativeai 
SpeechRecognition 
pydub 
python-dotenv